requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
marshmallow==3.20.1
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
validators==0.22.0
PyJWT==2.8.0
bcrypt==4.0.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
